refactor: 테스트 파일에서 CartPage 및 AdminPage의 경로 수정 및 useCoupons, useProducts 훅 추가
